Overhauled the Log Analytics clients APIs
