+++ src/SDK/MonitoringApplianceEmulator/App.config
@@ -12,16 +12,16 @@
     <!-- Auth extra query parameters to send to AAD for acquiring user tokens. -->
     <add key="AuthExtraQueryParameters" value="prompt=consent" />
     <!-- The format for creating Application Insights endpoint URI. -->
-    <add key="ApplicationInsightsUriFormat" value="https://api.applicationinsights.io/v1/apps/{0}/query" />
+    <add key="ApplicationInsightsQueryUriFormat" value="https://api.applicationinsights.io/v1/apps/{0}/query" />
     <!-- The format for creating Log Analytics endpoint URI. -->
-    <add key="LogAnalyticsUriFormat" value="https://api.loganalytics.io/v1/workspaces/{0}/query" />
+    <add key="LogAnalyticsQueryUriFormat" value="https://api.loganalytics.io/v1/workspaces/{0}/query" />
+    <!-- The format for creating resource-centric Log Analytics endpoint URI. -->
+    <add key="LogAnalyticsQueryUriFormat" value="https://api.loganalytics.io/v1{0}/query" />
     <!-- The URI of the resource manager. -->
     <add key="ResourceManagerBaseUri" value="https://management.azure.com/" />
     <!-- The enpoint of Log Analytics portal. -->
     <add key="LogAnalyticsPortalEndpoint" value="portal.loganalytics.io" />
     <!-- The enpoint of Application Insights portal. -->
     <add key="ApplicationInsightsPortalEndpoint" value="analytics.applicationinsights.io" />
-    <!-- The maximum number of telemetry resources that can be queried by the same query. -->
-    <add key="MaximumNumberOfTelemetryResources" value="300" />
   </appSettings>
 </configuration>
