+++ src/SDK/SmartDetectorsShared/Clients/AnalysisServicesFactory.cs
@@ -7,10 +7,8 @@
 namespace Microsoft.Azure.Monitoring.SmartDetectors.Clients
 {
     using System;
-    using System.Collections.Concurrent;
-    using System.Collections.Generic;
+    using System.Configuration;
     using System.Globalization;
-    using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
     using Microsoft.Azure.Monitoring.SmartDetectors;
@@ -24,8 +22,10 @@ namespace Microsoft.Azure.Monitoring.SmartDetectors.Clients
     /// </summary>
     public class AnalysisServicesFactory : IInternalAnalysisServicesFactory
     {
-        private const int MaxNumberOfResourcesInQuery = 10;
-        private readonly ConcurrentDictionary<string, IList<ResourceIdentifier>> subscriptionIdToWorkspaces = new ConcurrentDictionary<string, IList<ResourceIdentifier>>(StringComparer.CurrentCultureIgnoreCase);
+        private const string ApplicationInsightsQueryUriFormat = "https://api.applicationinsights.io/v1/apps/{0}/query";
+        private const string LogAnalyticsQueryUriFormat = "https://api.loganalytics.io/v1/workspaces/{0}/query";
+        private const string ResourceCentricQueryUriFormat = "https://api.loganalytics.io/v1{0}/query";
+
         private readonly ITracer tracer;
         private readonly IHttpClientWrapper httpClientWrapper;
         private readonly ICredentialsFactory credentialsFactory;
@@ -65,65 +65,43 @@ namespace Microsoft.Azure.Monitoring.SmartDetectors.Clients
         public bool UsedMetricClient { get; private set; }
 
         /// <summary>
-        /// Creates an instance of <see cref="ITelemetryDataClient"/>, used for running queries against data in log analytics workspaces.
+        /// Creates an instance of <see cref="ILogAnalyticsClient"/>, used for running Log Analytics queries on the specified resource. If the
+        /// resource type of <paramref name="resource"/> is <see cref="ResourceType.ApplicationInsights"/> then the created client will query
+        /// Application Insights telemetry.
         /// </summary>
-        /// <param name="resources">The list of resources to analyze.</param>
+        /// <param name="resource">The resource to analyze.</param>
         /// <param name="cancellationToken">The cancellation token</param>
-        /// <exception cref="TelemetryDataClientCreationException">A log analytics telemetry data client could not be created for the specified resources.</exception>
-        /// <returns>The telemetry data client, that can be used to run queries on log analytics workspaces.</returns>
-        public async Task<ITelemetryDataClient> CreateLogAnalyticsTelemetryDataClientAsync(IReadOnlyList<ResourceIdentifier> resources, CancellationToken cancellationToken)
+        /// <exception cref="TelemetryDataClientCreationException">A Log Analytics client could not be created for the specified resource.</exception>
+        /// <returns>The Log Analytics client, that can be used to run queries on <paramref name="resource"/>.</returns>
+        public async Task<ILogAnalyticsClient> CreateLogAnalyticsClientAsync(ResourceIdentifier resource, CancellationToken cancellationToken)
         {
-            // Verify that there are no Application Insights resources
-            if (resources.Any(resource => resource.ResourceType == ResourceType.ApplicationInsights))
-            {
-                throw new TelemetryDataClientCreationException($"Telemetry client creation failed - resources shouldn't contain resources of the type {ResourceType.ApplicationInsights}");
-            }
-
             // Mark that a log signal was used to create the alert
             this.UsedLogAnalysisClient = true;
 
-            // Get the resource IDs
-            IReadOnlyList<string> resourceIds = await this.GetResourceIdsForLogAnalyticsAsync(resources, cancellationToken);
-
-            // Create the client
-            return new LogAnalyticsTelemetryDataClient(this.tracer, this.httpClientWrapper, this.credentialsFactory, this.azureResourceManagerClient, resourceIds, this.queryTimeout);
-        }
-
-        /// <summary>
-        /// Creates an instance of <see cref="ITelemetryDataClient"/>, used for running queries against data in Application Insights.
-        /// </summary>
-        /// <param name="resources">The list of resources to analyze.</param>
-        /// <param name="cancellationToken">The cancellation token</param>
-        /// <exception cref="TelemetryDataClientCreationException">An Application Insights telemetry data client could not be created for the specified resources.</exception>
-        /// <returns>The telemetry data client, that can be used to run queries on Application Insights.</returns>
-        public Task<ITelemetryDataClient> CreateApplicationInsightsTelemetryDataClientAsync(IReadOnlyList<ResourceIdentifier> resources, CancellationToken cancellationToken)
-        {
-            // Verify that there are resources
-            if (!resources.Any())
+            // Split the flow according the resource type - Application Insights and Log Analytics workspace resources have their own query endpoints
+            string uriFormat;
+            string resourceId;
+            switch (resource.ResourceType)
             {
-                throw new TelemetryDataClientCreationException("No resources provided");
+                case ResourceType.ApplicationInsights:
+                    uriFormat = ConfigurationManager.AppSettings["ApplicationInsightsQueryUriFormat"] ?? ApplicationInsightsQueryUriFormat;
+                    resourceId = await this.azureResourceManagerClient.GetApplicationInsightsAppIdAsync(resource, cancellationToken);
+                    break;
+
+                case ResourceType.LogAnalytics:
+                    uriFormat = ConfigurationManager.AppSettings["LogAnalyticsQueryUriFormat"] ?? LogAnalyticsQueryUriFormat;
+                    resourceId = await this.azureResourceManagerClient.GetLogAnalyticsWorkspaceIdAsync(resource, cancellationToken);
+                    break;
+
+                default:
+                    uriFormat = ConfigurationManager.AppSettings["ResourceCentricQueryUriFormat"] ?? ResourceCentricQueryUriFormat;
+                    resourceId = resource.ToResourceId();
+                    break;
             }
 
-            // Verify that all resources are of type ApplicationInsights
-            if (resources.Any(resource => resource.ResourceType != ResourceType.ApplicationInsights))
-            {
-                throw new TelemetryDataClientCreationException($"Telemetry client creation failed - all resources must be of type {ResourceType.ApplicationInsights}");
-            }
-
-            // Verify there are not too many resources
-            if (resources.Count > MaxNumberOfResourcesInQuery)
-            {
-                throw new TelemetryDataClientCreationException($"Cannot run analysis on more than {MaxNumberOfResourcesInQuery} applications");
-            }
-
-            List<string> resourceIds = resources.Select(application => application.ToResourceId()).ToList();
-
-            // Mark that a log signal was used to create the alert
-            this.UsedLogAnalysisClient = true;
-
-            // Create the client
-            return Task.FromResult<ITelemetryDataClient>(new ApplicationInsightsTelemetryDataClient(
-                this.tracer, this.httpClientWrapper, this.credentialsFactory, this.azureResourceManagerClient, resourceIds, this.queryTimeout));
+            // And create the client
+            var queryUri = new Uri(string.Format(CultureInfo.InvariantCulture, uriFormat, resourceId));
+            return new LogAnalyticsClient(this.tracer, this.httpClientWrapper, this.credentialsFactory, queryUri, this.queryTimeout);
         }
 
         /// <summary>
@@ -160,60 +138,6 @@ namespace Microsoft.Azure.Monitoring.SmartDetectors.Clients
         {
             // Create the client
             return Task.FromResult<IActivityLogClient>(new ActivityLogClient(this.credentialsFactory, this.httpClientWrapper, this.tracer));
-        }
-
-        /// <summary>
-        /// Gets IDs of Log Analytics resources.
-        /// </summary>
-        /// <param name="resources">The resources</param>
-        /// <param name="cancellationToken">The cancellation token</param>
-        /// <returns>A <see cref="Task{TResult}"/>, returning the resource IDs</returns>
-        private async Task<IReadOnlyList<string>> GetResourceIdsForLogAnalyticsAsync(IReadOnlyList<ResourceIdentifier> resources, CancellationToken cancellationToken)
-        {
-            // Verify that there are resources
-            if (!resources.Any())
-            {
-                throw new TelemetryDataClientCreationException("No resources provided");
-            }
-
-            IReadOnlyList<ResourceIdentifier> workspaces;
-            if (resources.All(resource => resource.ResourceType == ResourceType.LogAnalytics))
-            {
-                // All resources are of type LogAnalytics. Create a client that queries all these workspaces.
-                workspaces = resources;
-            }
-            else
-            {
-                // Since we do not know where the telemetry of each resource is, create a client that queries all workspaces in the subscription.
-                var workspacesList = new List<ResourceIdentifier>();
-                foreach (string subscriptionId in resources.Select(resource => resource.SubscriptionId).Distinct(StringComparer.CurrentCultureIgnoreCase))
-                {
-                    // Try to get the workspaces list from the cache, and if it isn't there, use the Azure Resource Manager client
-                    IList<ResourceIdentifier> subscriptionWorkspaces;
-                    if (!this.subscriptionIdToWorkspaces.TryGetValue(subscriptionId, out subscriptionWorkspaces))
-                    {
-                        subscriptionWorkspaces = await this.azureResourceManagerClient.GetAllResourcesInSubscriptionAsync(subscriptionId, new[] { ResourceType.LogAnalytics }, cancellationToken);
-                        this.subscriptionIdToWorkspaces[subscriptionId] = subscriptionWorkspaces;
-                    }
-
-                    workspacesList.AddRange(subscriptionWorkspaces);
-                }
-
-                workspaces = workspacesList;
-                if (workspaces.Count == 0)
-                {
-                    throw new TelemetryDataClientCreationException("No log analytics workspaces were found");
-                }
-            }
-
-            // Verify there aren't too many resources
-            if (workspaces.Count > MaxNumberOfResourcesInQuery)
-            {
-                throw new TelemetryDataClientCreationException($"Cannot run analysis on more than {MaxNumberOfResourcesInQuery} applications");
-            }
-
-            List<string> workspacesResourceIds = workspaces.Select(workspace => workspace.ToResourceId()).ToList();
-            return workspacesResourceIds;
         }
     }
 }
